+++ previous-versions/files/src/codatfiles/models/operations/download_files.py
@@ -3,9 +3,6 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import schema as shared_schema
-from codatfiles import utils
-from dataclasses_json import Undefined, dataclass_json
 from typing import Optional
 
 
@@ -19,43 +16,15 @@ class DownloadFilesRequest:
 
 
 
-@dataclass_json(undefined=Undefined.EXCLUDE)
-@dataclasses.dataclass
-class DownloadFilesErrorMessage:
-    r"""One or more of the resources you referenced could not be found.
-    This might be because your company or data connection id is wrong, or was already deleted.
-    """
-    can_be_retried: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('canBeRetried'), 'exclude': lambda f: f is None }})
-    r"""`True` if the error occurred transiently and can be retried."""
-    correlation_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('correlationId'), 'exclude': lambda f: f is None }})
-    r"""Unique identifier used to propagate to all downstream services and determine the source of the error."""
-    detailed_error_code: Optional[int] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('detailedErrorCode'), 'exclude': lambda f: f is None }})
-    r"""Machine readable error code used to automate processes based on the code returned."""
-    error: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('error'), 'exclude': lambda f: f is None }})
-    r"""A brief description of the error."""
-    service: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('service'), 'exclude': lambda f: f is None }})
-    r"""Codat's service the returned the error."""
-    status_code: Optional[int] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('statusCode'), 'exclude': lambda f: f is None }})
-    r"""The HTTP status code returned by the error."""
-    
-
-
-
 @dataclasses.dataclass
 class DownloadFilesResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
     data: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Success"""
-    error_message: Optional[DownloadFilesErrorMessage] = dataclasses.field(default=None)
-    r"""One or more of the resources you referenced could not be found.
-    This might be because your company or data connection id is wrong, or was already deleted.
-    """
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""
-    schema: Optional[shared_schema.Schema] = dataclasses.field(default=None)
-    r"""The request made is not valid."""